fix: CI build process and Next.js onboarding page structure
- Add TypeScript dependency to root package.json
- Fix CI workflow to build packages in proper dependency order
- Fix JSON syntax error in services/api/package.json
- Fix Next.js 15 session handling with proper async cookies()
- Restructure onboarding page to separate server/client components
- All builds now pass successfully
