+++ apps/web/lib/session.ts
@@ -10,7 +10,7 @@ export type Session = {
 export async function getSession(): Promise<Session> {
   // Next 15 App Router: use headers() rather than cookies() on the server to avoid edge/runtime quirks
   const { cookies } = await import("next/headers");
-  const jar = cookies();
+  const jar = await cookies();
   const raw = jar.get("__session")?.value;
   if (!raw) return null;
   try {
