+++ package.json
@@ -15,6 +15,7 @@
     "@typescript-eslint/eslint-plugin": "^8.42.0",
     "@typescript-eslint/parser": "^8.42.0",
     "eslint": "^9.35.0",
+    "typescript": "^5.5.4",
     "typedoc": "^0.28.12"
   }
 }
