+++ services/api/package.json
@@ -7,8 +7,8 @@
     "dev": "tsx watch src/index.ts",
     "build": "tsc -p tsconfig.json",
     "start": "node dist/index.js",
-  "typecheck": "tsc -p tsconfig.json --noEmit",
-  "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\""
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\""
   },
   "dependencies": {
     "cors": "2.8.5",
